Add option for alternative home
Alternative home is suitable for cases that the PolarBearBlog is only
for blog and there's a separated domain for the home page. When the
alternative home URL is set, "/" is routed to index page instead of the
home page and the "Home" link is linking to the external site instead.